Add `getWebStream()` for streaming responses to browsers (#905)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Introduces a new `getWebSteam()` method to use when passing the stream
back as a `Response` to a browser.

This method is equivalent to the `getStream()` method, but will
correctly encode data to be compliant with SSE.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Undocumented atm, likely coming following #906
- [ ] ~Added unit/integration tests~ N/A No suites for this yet
- [x] Added changesets if applicable

## Related

- Fixes for #906

---------

Co-authored-by: Charly Poly <1252066+charlypoly@users.noreply.github.com>