When (de)serializing errors, allow `unknown` `cause`, but still parse (#922)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

`Error.cause` should be `unknown`, though we still wish to _attempt_ to
parse it as an `Error` so that we can recursively expand causes as they
would be before JSON (de)serialization.

These changes loosen `serializeError` and `deserializeError` to allow
recursively expanding `cause` and falling back to `unknown`.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [ ] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Complements inngest/inngest#2088