Add dependency injection middleware (#919)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Adds `dependencyInjectionMiddleware()`, allowing users to easily add to
function input without needing to understand middleware.

```ts
import { dependencyInjectionMiddleware } from "inngest";

const prisma = new Prisma()

const inngest = new Inngest({
  id: 'my-app',
  middleware: [dependencyInjectionMiddleware({ prisma })],
});
```

Note that due to TypeScript constraints we can't use `const` for the
generic. Following #903 and #920 we can do this.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [x] Added a [docs PR](https://github.com/inngest/website) that
references this PR
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- Documented in inngest/website#1090
- Functionality hindered by TS support, solved by:
  - #903 
  - #920