Fix connecting to prod realtime (#927)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Fix bug where WebSocket connections do not allow being redirected, as is
the case when hitting the prod API with `ws://`, causing a `301`.

Vercel's edge runtime appropriately complies with the WebSocket spec
here, whereas their local `dev` doesn't, hence the discrepancy.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [ ] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Complements inngest/inngest#2292