Fix validation middleware silently failing (#953)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Removes the `instanceof` check when checking if a given runtime schema
is compatible with Zod; this will not return truthy if the schema passed
was created with a different version of Zod, which can happen if
dependencies of an app both depend on different versions of Zod that are
not compatible.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [ ] ~Added unit/integration tests~ N/A Multi-package; covered
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Fixes #932