Do not dynamically construct env vars in a browser env (#959)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

In browser envs, do not dynamically construct env vars, as they likely
inline env vars.

A potential partial fix for #957, which also includes finding the env
vars across a couple of other runtimes.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Fix
- [ ] ~Added unit/integration tests~ N/A Covered
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Fixes #957