Add ability for signal waits to supersede others (#985)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Add the ability for `step.waitForSignal()` to overwrite any other
existing waits in any run in the environment.

```ts
await step.waitForSignal("step-id", {
  signal: "my-signal",
  timeout: "5m",
  onConflict: "replace",
});
```

Also requires a new option for every `step.waitForSignal()` call:
`onConflict`.
- `onConflict: "replace"` will replace any existing wait with this one,
leaving the replaced wait to time out
- `onConflict: "fail"` will fail this run if another wait for the same
signal exists - this was the previous default behaviour

This is a required option for now to make it so it's a considered part
of every wait. It may fall back to be optional with a default later.



## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A
- [ ] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Requires inngest/inngest#2449