Export types for convenience and discoverability (#997)
## Summary
With all types exported on the `"inngest/connect"` path along side the
main `connect()` function, users can more easily click into this file
and discover related types they may want to use. This is helpful
compared to previously having to dig into the code or view the docs to
know where these types are located. This also makes import statements
simpler for end users as they can have all related imports on a single
line.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] Added a [docs PR](https://github.com/inngest/website) that
references this PR
- [ ] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- INN-5009