Revert "obscure "node:async_hooks" import from bundlers" (#1028)
The inngest/inngest-js#1020 PR introduced breaking compatibility changes
with Next.js, resulting in numerous bug reports in the past days.

The new code block raises the following error: 

```
Error: Cannot find module as expression is too dynamic
    at async (../../../src/components/execution/als.ts:63:12) {
  code: 'MODULE_NOT_FOUND'
}
```

resulting in the disablement of the ALS feature required for AgentKit to
function in an Next.js project (and others):

```
node:async_hooks is not supported in this runtime. Experimental async context is disabled.
```