SYS-285: Add support for conditional batching (#1050)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Add `if` support in the batch config. This conditional boolean
expression, when provided, would determine which events get batched
together for execution.

If the expression fails to evaluate or if it evaluates to false, the
event will be treated ineligible for batching and will be scheduled for
execution immediately.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [x] Added a [docs PR](https://github.com/inngest/website/pull/1232)
that references this PR
- [ ] ~Added unit/integration tests~ N/A
- [x] Added changesets if applicable

## Related
- SYS-283
- OSS feature PR: https://github.com/inngest/inngest/pull/2818
- Documentation PR: https://github.com/inngest/website/pull/1232