Use `Symbol.toStringTag` for `*.Like` types (#924)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Use
[`Symbol.toStringTag`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol/toStringTag)
for `*.Like` interfaces, allowing much easier cross-version checks.

If many different versions of `inngest` are installed in a user's
package (commonly due to `inngest` being a sub-dependency with differing
range specifiers), comparing types can be tricky. A `serve()` call may
accept an array of `InngestFunction`s, but that class across two
versions is not equivalent, even if the code has not changed at all.


[_Playground_](https://www.typescriptlang.org/play/?#code/MYGwhgzhAECSB28DmBTCAXAagRmgbwChpoAHAJwEsA3MdFaCgE2gF5oAiAMwHtv2BuAgF8CBUJBgJkaLACZ8RUpRp0GzNl14Dho4N3gZoZAK7xW0ABQV4JY+gBccRKgw4AlKwB80U4xSdrFEZBAhN4C3gUAHcnaVdZCzc3QQB6FOgAQTIkYwBbFHh0aG5OaHQATxJ6AHIpFzlqhhh4biKJCiR4MAAjEHp0blIwMjB8ujJi0oqq6FrnGRxqgDoCNOJoABVKtGgACzAqeggUEmHaej9xEfQKfRgS6DAlanOlbiqyCtmmZYtZAGYACwAVjcQA)

```ts
class InngestV1 {
  private id = "foo";
}

class InngestV2 {
  private id = "foo";
}

const run = (input: InngestV1) => undefined;

run(new InngestV2());
// Argument of type 'InngestV2' is not assignable to parameter of type 'InngestV1'.
//   Types have separate declarations of a private property 'id'.(2345)
```

This is especially prevelant in the immediate user-facing API and when
passing `Inngest` apps, `InngestFunction`s, or `InngestMiddleware` into
various calls.

In #817 we added `*.Like` types such as `Inngest.Like`, which are
simplified summaries of the public-facing contract for that object.

```ts
export namespace Inngest {
  export interface Like {
    readonly id: string;
    apiBaseUrl: string | undefined;
    eventBaseUrl: string | undefined;
    env: string | null;
    buildId?: string | undefined;
  }
}
```

And we make sure that the classes implement these `Like` interfaces so
we don't break the contract:

```ts
export class Inngest implements Inngest.Like {
  // ...
}
```

However, this prevents us from iterating on inference and means that all
future values must also be optional. See here how we're restricted from
inferring an `Inngest` app's ID:


[_Playground_](https://www.typescriptlang.org/play/?#code/JYOwLgpgTgZghgYwgAgDLANYoN4ChnLAAmAXMgM5hSgDmA3LgL664IA2c55yAkiCDQiUAagEZCAWwAObCBIjhu6LMjwEpAVwBGbYAkKkKVWsgC8yAEQSAngFo4UqRYbNWHLr36CRAJgA8CAD2IJTIACo8RMgQAB6QIETclNQCAHySMnIKYEqYOPjImjp6BmQRUeYWAO6BVc4FAPQN4dZSKADkybTthNwggWDInOTANCBwOihggchgrR3l7QB0jc0EncYCPcDcw6Pjk7MzYAAWKEEhVHCgg4Ews-PI7YsANMhaGoPPkT1BGmxRLQoUCUODgYBwSBRKrAU5DZBEYAwGDQbIUbRzNrIO7IC7Ja7gJ5dLZLAAUPgAzD4fABKJhAA)

```ts
interface Like {
  id: string;
}

class InngestV1 implements Like {
  public id: string = "my-app";
}

class InngestV2<const TId extends string> implements Like {
  public id: TId = "wow";
  // Type 'string' is not assignable to type 'TId'.
  //   'string' is assignable to the constraint of type 'TId', but 'TId' could be instantiated with a different subtype of constraint 'string'.(2322)
}
```

The only time we care about this is internally, for us to understand
that we've been given something `Inngest*`ish. For many cases, this also
means that we want to access private methods on the object which the
`Like` interface can't provide us information about anyway.

Therefore we come to
[`Symbol.toStringTag`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol/toStringTag).
The upcoming
[`Temporal`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Temporal)
object uses this to mark its various related objects (see #918) and for
us it makes checking for an `Inngest`ish entity really easy.

```ts
// Check for a particular type
(input as any)[Symbol.toStringTag] === "Inngest.App";

// Check for anything to do with Inngest
(input as any)[Symbol.toStringTag]?.startsWith?.("Inngest.");
```

This means we can easily check at the type level and at runtime if the
entity is what we expect and act accordingly. Additionally, stringifying
the object (in logs or errors, for instance) is much clearer about what
we're seeing:


[_Playground_](https://www.typescriptlang.org/play/?#code/MYGwhgzhAECSB28DmBTCAXA6gS3QCwHsBXdAFTCWgG8AoaaAByICMRthpsATALmgwBO2ZNAC80AEQBbAJ4BaMAwYSA3DQC+NGqEgwEyNFlx5ylWvVTpoAbQDKMqcwIgAdOgK30Q5KYC6ACgBKajp6aAEUdCIBeEl9VAwXAEElVVDNTW0CeAhnFBcQAiR-AAN4wxx8YjIKaAASKngUAHc4RASjKpJTIPUSwLVgbNyQfMLisvaK41N6xpa2gwxKkwpe-rUgA)

```ts
class InngestWithoutTag {
  public id: string = "my-app";
}

class InngestWithTag {
  get [Symbol.toStringTag]() {
    return "Inngest.App";
  }
}

console.log(`InngestWithoutTag ${new InngestWithoutTag()}`);
console.log(`InngestWithTag ${new InngestWithTag()}`);
```

```
InngestWithoutTag [object Object]
InngestWithTag [object Inngest.App]
```

In practice, internals still need to be careful about accessing private
properties, but this is a problem that exists with the `Like` method
too; it's just something we should be aware of when dealing with a
`Inngest`ish entity passed through a public API.

When we make this change, we must ensure that all of our related
packages also update their required `inngest` versions in order to allow
the typing to seamlessly move across with a couple of upgrades.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A KTLO
- [ ] ~Added unit/integration tests~ N/A Covered
- [x] Added changesets if applicable

## Related

- Follows #817
- Inspiration from #918