Add support for Standard Schemas (#1055)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Instead of attempting to directly support both Zod v3 and v4 together,
this PR introduces an `EventSchemas#fromSchema()` method, which allows
any schema to be specified so long as it's compliant as a [Standard
Schema](https://standardschema.dev/).

This means we deprecate `.fromZod()` and encourage use of
`.fromSchema()` instead, as this supports `zod@>=3.40.0` and a multitude
of other validation libraries.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Documenting in #920
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Targeted for release with #920
- Addresses #933