Freeze lockfiles in CI for pnpm (#1060)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

pnpm, like npm and Yarn, all use `--frozen-lockfile` (or their
equivalent) when in CI environments. Still, adding this to be very
explicit about it.

Also includes a small audit check in each PR.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A KTLO
- [ ] ~Added unit/integration tests~ N/A KTLO
- [x] Added changesets if applicable