TanStack pinning (#1099)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

TanStack CI was getting a big angry and seemed to be drifting to a new
minor version (as is expected) which unfortunately had a breaking import
change.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A KTLO
- [ ] ~Added unit/integration tests~ N/A Covered - making this green
- [x] Added changesets if applicable