Weekly flake bumps (#1104)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Adding an automated PR to bump the flake lock weekly.

Just helps us keep up to date with the latest version we're looking for
of e.g. Node.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Internal
- [x] Added unit/integration tests
- [x] Added changesets if applicable