Fix `.fromSchema()` runtime schema being mismatched (#1114)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Fixes a bug whereby the runtime schemas being added when using
`.fromSchema()` were not compatible with
`@inngest/middleware-validation`.

`@inngest/middleware-validation` tests are expected to fail, as they use
the latest published `inngest`.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [x] Added unit/integration tests
- [x] Added changesets if applicable