Export types for the step result (#1129)
## Summary

#1111 removed direct access to the files/helpers. `Jsonify` is useful
when one need to pass step output to some helper methods. It's quite
hard to type that helper method params without wrapping existing type
with the conversion that inngest does to serialize step result.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~
- [ ] ~~Added unit/integration tests~~
- [x] Added changesets if applicable

---------

Co-authored-by: MasterLambaster <MasterLambaster@gmail.com>