Support for latest OpenAI responses API (#1134)
## Summary
Created model and adapter for the latest OpenAI responses API.

We currently only support the OpenAI chat API which is now considered to
be deprecated. This PR introduces a new model called `openaiResponses`
which can be implemented like this:

```typescript
import { openaiResponses } from "@inngest/ai";
import { inngest } from "./client";

export const runTest = inngest.createFunction(
  {
    id: "run-openai-test",
    name: "OpenAI Responses Test",
  },
  { event: "agent/chat.requested" },
  async ({ step }) => {
    try {
      const result = await step.ai.infer("test-openai", {
        model: openaiResponses({
          model: "gpt-5",
          defaultParameters: { reasoning: { effort: "medium" } },
        }),
        body: {
          input: [
            {
              role: "user",
              content:
                "Hello! Can you explain to me how redux state management works for React?",
            },
          ],
        },
      });

      return {
        success: true,
        result: result.output_text,
      };
    } catch (error) {
      console.error(error);
      throw error;
    }
  }
);

```

I've also created a few unit and smoke tests. Make sure to add a .env
file to the AI package with an `OPENAI_API_KEY` to run the smoke tests.
To run the tests, add your env vars and run: `pnpm test` and `pnpm
test:smoke`.

This will run tests for both gemini and openai (gpt-5) models.

Once this is merged in, we can work on adapters to integrate this into
AgentKit as well.



## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] Added a [docs PR](https://github.com/inngest/website) that
references this PR
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

[OpenAI Responses API
Docs](https://platform.openai.com/docs/api-reference/responses)

[Issue #204 in
AgentKit](https://github.com/inngest/agent-kit/issues/204)