Re-export realtime types from `middleware` entrypoint (#1166)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Re-exports `./types` from the `@inngest/realtime/middleware` entrypoint.

This is already exported from the main `@inngest/realtime` entrypoint,
but this is no longer used for the middleware specifically.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~Added a [docs PR](https://github.com/inngest/website) that
references this PR~ N/A Bug fix
- [ ] ~Added unit/integration tests~ N/A but needs a composite test
- [x] Added changesets if applicable