Rewrite test for steps capturing input/variables
The test as written doesn't test anything because capturedStep1Result
was being mocked directly in the mock step itself. Although the intent
is to show that `step1Result` is captured, this isn't possible with mocks
because JS is lexically scoped and so the mock step can't touch a
variable inside the original Inngest function even if the non mocked
step can.

However, mocked steps can reference step inputs, so I rewrote the test
to show a variety of cases with mock/real steps using inputs.

2 of the tests fail as of this commit